feat(epic-42): Story 42.7 — Pydantic contract tightening, extra=forbid, mypy zero
- All 16 models in pipeline_context.py migrated from extra=allow to extra=forbid
- Added explicit fields to TextBlock (is_mono, color, sub_spans), ImageInfo
  (bbox_valid, format), DocumentTreeNode (30+ node-type-specific fields)
- PipelineResult expanded with all result_assembly.py fields; PipelineResultResponse
  wrapper created and declared as response_model in GET /analyze/{job_id}/result
- mypy: 0 errors in 72 source files (fixed no-any-return, unused-ignore,
  arg-type, attr-defined errors across 10+ files)
- Resolved pre-existing merge conflicts in main.py, job_store.py,
  stage1_layout_clustering.py, supabase_gateway.py from stash pop
- 448 tests passing; 1 pre-existing failure confirmed

+++ backend/services/stages/stage1_clustering/clustering_algorithms.py
@@ -105,7 +105,7 @@ def _density_similarity(
         total_area = sum(
             (b["bbox_norm"][2] - b["bbox_norm"][0]) * (b["bbox_norm"][3] - b["bbox_norm"][1]) for b in blocks
         )
-        return total_area / body_height
+        return float(total_area / body_height)
 
     d_a = _compute_density(core_a)
     d_b = _compute_density(core_b)
